# Conflicts:
#	src/main/java/com/wangbei/BuddhaApplication.java
#	src/main/java/com/wangbei/awre/ApplicationContextBeanFactory.java
#	src/main/resources/application.properties
#	src/main/resources/templates/index.html
#	src/main/resources/templates/login.html

+++ src/main/java/com/wangbei/dao/impl/jpa/MenuRepository.java
@@ -9,9 +9,9 @@ import com.wangbei.entity.Menu;
 * @class com.wangbei.dao.impl.jpa.MenuRepository
 * @description 菜单jpa接口
 */
-public interface MenuRepository extends Repository<Menu,Long> {
+public interface MenuRepository extends Repository<Menu,Integer> {
 
-    Menu findById(Long id);
+    Menu findById(Integer id);
 
     Menu save(Menu menu);
